+++ Cargo.toml
@@ -27,6 +27,11 @@ members = [
 	"frame/dvm/rpc/runtime-api",
 	"frame/elections-phragmen",
 	"frame/evm",
+	"frame/evm/precompile/simple",
+	"frame/evm/precompile/modexp",
+	"frame/evm/precompile/ed25519",
+	"frame/evm/precompile/bn128",
+	"frame/evm/precompile/blake2",
 	"frame/header-mmr",
 	"frame/header-mmr/rpc",
 	"frame/header-mmr/rpc/runtime-api",
