+++ bin/node/runtime/pangolin/Cargo.toml
@@ -26,6 +26,7 @@ darwinia-elections-phragmen         = { default-features = false, path = "../../
 darwinia-ethereum-backing           = { default-features = false, path = "../../../../frame/bridge/ethereum/backing" }
 darwinia-ethereum-relay             = { default-features = false, path = "../../../../frame/bridge/ethereum/relay" }
 darwinia-evm                        = { default-features = false, path = "../../../../frame/evm" }
+pallet-evm-precompile-simple        = { default-features = false, path = "../../../../frame/evm/precompile/simple" }
 darwinia-header-mmr                 = { default-features = false, path = "../../../../frame/header-mmr" }
 darwinia-header-mmr-rpc-runtime-api = { default-features = false, path = "../../../../frame/header-mmr/rpc/runtime-api" }
 darwinia-relay-authorities          = { default-features = false, path = "../../../../frame/bridge/relay-authorities" }
@@ -98,6 +99,7 @@ std     = [
 	"darwinia-ethereum-backing/std",
 	"darwinia-ethereum-relay/std",
 	"darwinia-evm/std",
+	"pallet-evm-precompile-simple/std",
 	"darwinia-header-mmr/std",
 	"darwinia-header-mmr-rpc-runtime-api/std",
 	"darwinia-relay-authorities/std",
