+++ frame/evm/Cargo.toml
@@ -15,16 +15,10 @@ codec                 = { package = "parity-scale-codec", version = "1.3.5", def
 evm                   = { version = "0.19.0", default-features = false, features = ["with-codec"] }
 evm-gasometer         = { version = "0.19.0", default-features = false }
 evm-runtime           = { version = "0.19.0", default-features = false }
-impl-trait-for-tuples = { version = "0.1" }
 primitive-types       = { version = "0.7.0", default-features = false, features = ["rlp", "byteorder"] }
-ripemd160             = { version = "0.9", default-features = false }
 rlp                   = { version = "0.4", default-features = false }
 serde                 = { version = "1.0.118", optional = true, features = ["derive"] }
 sha3                  = { version = "0.8", default-features = false }
-num 			      = { version = "0.3", features = ["alloc"], default-features = false, optional = true }
-ethereum-types        = { version = "0.9.2", default-features = false, optional = true }
-bn = { package        = "substrate-bn", version = "0.5", default-features = false, optional = true }
-ed25519-dalek         = { version = "1.0.0", features = ["alloc", "u64_backend"], default-features = false, optional = true }
 # darwinia
 darwinia-balances       = { default-features = false, path = "../balances" }
 darwinia-evm-primitives = { default-features = false, path = "../../primitives/evm" }
@@ -41,11 +35,7 @@ sp-std           = { default-features = false, git = "https://github.com/darwini
 darwinia-support = { features = ["easy-testing"], path = "../support" }
 
 [features]
-default = ["std", "modexp", "ed25519", "bn128", "blake2f"]
-blake2f = []
-modexp = ["num"]
-ed25519 = ["ed25519-dalek"]
-bn128 = ["bn", "ethereum-types"]
+default = ["std"]
 
 std = [
 	"crates-std",
@@ -70,7 +60,6 @@ substrate-std = [
 	"frame-system/std",
 	"pallet-timestamp/std",
 	"primitive-types/std",
-	"ripemd160/std",
 	"rlp/std",
 	"sha3/std",
 	"sp-core/std",
