+++ frame/evm/precompile/blake2/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "pallet-evm-precompile-blake2"
+version = "2.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+license = "Apache-2.0"
+homepage = "https://substrate.dev"
+repository = "https://github.com/paritytech/substrate/"
+description = "BLAKE2 precompiles for EVM pallet."
+
+[dependencies]
+evm                     = { version = "0.19.0", default-features = false, features = ["with-codec"] }
+darwinia-evm-primitives = { default-features = false, path = "../../../../primitives/evm" }
+sp-core                 = { default-features = false, git = "https://github.com/darwinia-network/substrate.git", branch = "common-library" }
+sp-io                   = { default-features = false, git = "https://github.com/darwinia-network/substrate.git", branch = "common-library" }
+
+[features]
+default = ["std"]
+
+std = [
+	"evm/std",
+	"darwinia-evm-primitives/std",
+	"sp-core/std",
+	"sp-io/std",
+]
