+++ frame/evm/precompile/bn128/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "pallet-evm-precompile-bn128"
+version = "2.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+license = "Apache-2.0"
+homepage = "https://substrate.dev"
+repository = "https://github.com/paritytech/substrate/"
+description = "BN128 precompiles for EVM pallet."
+
+[dependencies]
+evm						 = { version = "0.19.0", default-features = false, features = ["with-codec"] }
+bn						 = { package = "substrate-bn", version = "0.5", default-features = false }
+sp-core                  = { default-features = false, git = "https://github.com/darwinia-network/substrate.git", branch = "common-library" }
+sp-io                    = { default-features = false, git = "https://github.com/darwinia-network/substrate.git", branch = "common-library" }
+darwinia-evm-primitives  = { default-features = false, path = "../../../../primitives/evm" }
+
+[features]
+default = ["std"]
+
+std = [
+	"evm/std",
+	"darwinia-evm-primitives/std",
+	"sp-core/std",
+	"sp-io/std",
+]
