+++ frame/evm/precompile/modexp/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "pallet-evm-precompile-modexp"
+version = "2.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+license = "Apache-2.0"
+homepage = "https://substrate.dev"
+repository = "https://github.com/paritytech/substrate/"
+description = "MODEXP precompiles for EVM pallet."
+
+[dependencies]
+sp-core                  = { default-features = false, git = "https://github.com/darwinia-network/substrate.git", branch = "common-library" }
+sp-io                    = { default-features = false, git = "https://github.com/darwinia-network/substrate.git", branch = "common-library" }
+darwinia-evm-primitives = { default-features = false, path = "../../../../primitives/evm" }
+evm = { version = "0.19.0", default-features = false, features = ["with-codec"] }
+num = { version = "0.3", features = ["alloc"], default-features = false }
+
+[features]
+default = ["std"]
+std = [
+	"sp-core/std",
+	"sp-io/std",
+	"darwinia-evm-primitives/std",
+	"evm/std",
+	"num/std",
+]
