+++ primitives/evm/Cargo.toml
@@ -14,6 +14,7 @@ version     = "1.3.0"
 codec = { package = "parity-scale-codec", version = "1.3.5", default-features = false }
 evm   = { version = "0.19.0", default-features = false, features = ["with-codec"] }
 serde = { version = "1.0.118", optional = true, features = ["derive"] }
+impl-trait-for-tuples = "0.1"
 # substrate
 sp-core = { default-features = false, git = "https://github.com/darwinia-network/substrate.git", branch = "common-library" }
 sp-std  = { default-features = false, git = "https://github.com/darwinia-network/substrate.git", branch = "common-library" }
